% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptm_adjustment_by_global_2.R
\name{adjust_for_parent_protein_profiles_2}
\alias{adjust_for_parent_protein_profiles_2}
\title{Adjusting PTM profiles by global. Version 2}
\usage{
adjust_for_parent_protein_profiles_2(m_ptm, m_global, method = "subtraction")
}
\arguments{
\item{m_ptm}{MSnSet with PTM data}

\item{m_global}{MSnSet with global data}

\item{method}{One of "subtraction", "compensation" or "regression"}
}
\value{
MSnSet object
}
\description{
Adjusting PTM relative abundance profiles by the global profiles of the
parent protein. The motivation behind this adjustment is that there are two
contributions to the observed PTM abundance changes - change in the
PTM stoichiometry and the change in the parent protein abundance.
}
\details{
\itemize{
   \item \bold{subtraction:} This simple method works best if there is
   no dynamic range compression. Otherwise, if global data has higher
   compression, the PTM data will be undercorrected. It calculates
   the adjusted profile as: \eqn{PTM_{adj} = PTM_{obs} - Protein_{global}}.

   \item \bold{regression:} This method fits a linear model for each feature:
   \eqn{PTM_{obs} \sim \beta_0 + \beta_1 \cdot Protein_{global}}. Then
   \eqn{PTM_{adj} = PTM_{obs} - \beta_0 + \beta_1 \cdot Protein_{global}}
   Works well only if there are lots of samples with varying change
   (e.g. cohort of human subjects). It probably won't work for A vs B
   comparison. Since this is basically two points, there is no
   room for regression.

   \item \bold{moderation:} DO NOT USE! The method is similar to `regression`.
   However, the \eqn{\beta_1} coefficients are modeled as a mixture of two
   normal distributions using the Expectation-Maximization approach.
   One, with smaller standard deviation is assumed
   composed of PTMs with non-changing stoichiometry and the \eqn{\beta_1}
   coefficient, probably less than 1, reflects the dynamic range compression.
   The other distribution, with larger standard deviation reflects PTMs with
   truly changing stoichiometry. Further, the probabilities of belonging to
   non-changing and changing distributions are derived. If \eqn{p} is the
   probability that the site is changing, then the coefficient
   is \eqn{\hat{\beta}_1 = p \cdot \mu_{nonchanging} + (1-p) \cdot \beta_1}.
   \eqn{PTM_{adj} = PTM_{obs} - \beta_0 + \hat{\beta}_1 \cdot Protein_{global}}

   \item \bold{compensation:} This algorithm can be viewed as a continuation:
   `regression` -> `moderation` -> `compensation`. Essentially this is an
   extreme moderation. That is, all the \eqn{\beta_1} are shrunk on just one value -
   the mean of the narrower (smaller standard deviation) distribution (modeled via the global median).
   \eqn{PTM_{adj} = PTM_{obs} - \mu_{nonchanging} \cdot Protein_{global}}
}

Features that do not have matching parent protein data or have insufficient
data points for regression are returned as \code{NA} and subsequently
filtered.
}
