% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_stats_utils.R
\name{affinity_enrichment_stats}
\alias{affinity_enrichment_stats}
\title{Affinity Enrichment Statistics}
\usage{
affinity_enrichment_stats(m1, m2, prefix = NULL)
}
\arguments{
\item{m1}{MSnSet with enriched data. Replicates will be averaged.}

\item{m2}{MSnSet with non-enriched data. Replicates will be averaged.}

\item{prefix}{appends this string to the column names of the output
data.frame}
}
\value{
A data.frame containing:
\itemize{
  \item \code{diff}: log2(Enriched/Non-Enriched)
  \item \code{pval}: Probability of observing the ratio under the null normal distribution
  \item \code{padj}: Benjamini-Hochberg adjusted p-value
}
}
\description{
Models log2(Enriched/Non-Enriched) ratio as normal distribution.
The computes the p-values that the a particular point is part of the
distribution.
}
