% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC_report.R
\name{create_boxplot}
\alias{create_boxplot}
\title{Create Sample Boxplots}
\usage{
create_boxplot(msnset, fill_by, facet_by)
}
\arguments{
\item{msnset}{MSnSet. The object containing expression data.}

\item{fill_by}{Character. The pData column name for color-coding boxes.}

\item{facet_by}{Character. The pData column name for faceting the plot.}
}
\value{
A ggplot object.
}
\description{
Generates boxplots of expression distributions across samples,
faceted and colored by pData columns.
}
