% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC_report.R
\name{create_faceted_density_plot}
\alias{create_faceted_density_plot}
\title{Create Faceted Density Ridge Plot}
\usage{
create_faceted_density_plot(msnset, facet_by, fill_by)
}
\arguments{
\item{msnset}{MSnSet. The object containing expression data.}

\item{facet_by}{Character. The pData column to group plots by.}

\item{fill_by}{Character. The pData column used to color the ridges.}
}
\value{
A ggplot object.
}
\description{
Generates density ridge plots for each sample, chunked into faceted groups.
}
\details{
Tip: to get a consistent/desired font, add to the output of this function the following:
\code{ggplot2::theme(
         text = element_text(family = "Font Name"),
         axis.text.x = element_text(family = "Font Name", size = 4
      )}
}
