% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC_report.R
\name{create_layered_density_plot}
\alias{create_layered_density_plot}
\title{Create Layered Density Plot}
\usage{
create_layered_density_plot(msnset, color_by)
}
\arguments{
\item{msnset}{MSnSet. The object containing expression data.}

\item{color_by}{Character. The pData column used to color the density lines.}
}
\value{
A ggplot object.
}
\description{
Overlays density distribution curves for all samples on a single set of axes.
}
