% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC_report.R
\name{create_missingness}
\alias{create_missingness}
\title{Create Missingness Plot}
\usage{
create_missingness(
  msnset,
  font_family = "Courier New",
  feature_kind = "Feature"
)
}
\arguments{
\item{msnset}{MSnSet. The object containing expression data.}

\item{font_family}{Character. The font family to use for text elements.}

\item{feature_kind}{Character. Label for the type of features (e.g., "Feature").
Only used for legend labeling.}
}
\value{
A ggplot object.
}
\description{
Generates a visualization of missing values across samples using
\code{MSnbase::plotNA} as an engine.
}
