% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_processing.R
\name{create_msnset}
\alias{create_msnset}
\title{Create MSnSet Object from Ratios and FASTA}
\usage{
create_msnset(
  path_to_ratios_or_ratios,
  path_to_fasta_for_f_data,
  p_data_df,
  feature_name_col = c("ensembl_protein", "ensembl_gene"),
  semicolon_warning = TRUE,
  subset_samples_by_p_data = TRUE,
  save_as_rds_path = NULL
)
}
\arguments{
\item{path_to_ratios_or_ratios}{Character or data.frame. Path to a file containing ratios or a data.frame of ratios.}

\item{path_to_fasta_for_f_data}{Character. Path to the FASTA file used for feature data.}

\item{p_data_df}{data.frame. Metadata with row names matching sample names in ratios.}
}
\value{
An MSnSet object containing the processed data.
}
\description{
Generates an MSnSet object using ratio data and a FASTA file for feature data.
}
