% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC_report.R
\name{create_overall_correlation_heatmap}
\alias{create_overall_correlation_heatmap}
\title{Create Overall Sample Correlation Heatmap}
\usage{
create_overall_correlation_heatmap(msnset, annotation_by, width, height)
}
\arguments{
\item{msnset}{MSnSet. The object containing expression data.}

\item{annotation_by}{Character. The pData column name used to annotate samples.}

\item{width}{unit. The width of the heatmap.}

\item{height}{unit. The height of the heatmap.}
}
\value{
A HeatmapList object, returned invisibly after drawing.
}
\description{
Calculates and visualizes Pearson correlation between all sample pairs
in an MSnSet.
}
\details{
This function requires width and height to be specified because
the underlying plotting technology (ComplexHeatmap) is best used with explicit sizing.
}
