% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC_report.R
\name{.get_density_df}
\alias{.get_density_df}
\title{Prepare Data for Density Plots}
\usage{
.get_density_df(msnset, facet_by, fill_by, do_paste = TRUE)
}
\arguments{
\item{msnset}{MSnSet. The object containing expression data and pData.}

\item{facet_by}{Character. pData column name used for faceting.}

\item{fill_by}{Character. pData column name used for fill colors.}

\item{do_paste}{Logical. Whether to append the column name \code{facet_by}
 to its values. For example, if \code{facet_by = "tmt_plex"} and a value is
\code{"Plex1"}, it becomes \code{"tmt_plex Plex1"}. Default is TRUE.}
}
\value{
A data.frame in long format.
}
\description{
Reshapes an \code{MSnSet}'s exprs data into a long-format data frame
("long-format" being "each row is a single expression value."), suitable for density
and ridge plots using \code{ggplot2}.
}
