% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_exporting.R
\name{export_exprs}
\alias{export_exprs}
\title{Export Expression Data from MSnSet}
\usage{
export_exprs(m, filename, arrange_by = c())
}
\arguments{
\item{m}{An \code{MSnSet} object.}

\item{filename}{Character string. The path to the output CSV file.}

\item{arrange_by}{Character vector. Variable names in \code{pData(m)} to sort the columns by.}
}
\value{
No return value, called for side effect of writing a file.
}
\description{
Exports the expression data (\code{exprs(m)}) from an \code{MSnSet} object to a CSV file.
Sample names are cleaned (removing leading "X" and replacing "_" with " ") and
columns can be reordered based on \code{pData} variables.
}
