% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_processing.R
\name{organize_mzml_by_plex}
\alias{organize_mzml_by_plex}
\title{Organize mzML Files by Plex}
\usage{
organize_mzml_by_plex(
  src_mzML_dir,
  dest_mzML_subdirs = "mzML/plexes/",
  overwrite = FALSE,
  sub_with_cap_grp_1 = "^(\\\\d{2}).*"
)
}
\arguments{
\item{src_mzML_dir}{Character. Base directory containing the mzML folder.}

\item{dest_mzML_subdirs}{Character. Name of mzML subdirectory. Default is "plexes".}

\item{overwrite}{Logical. Whether to overwrite existing files. Default is FALSE.}

\item{sub_with_cap_grp_1}{Character. Used to obtain plex ID through a
call to `sub(sub_with_cap_grp_1, "\\1", <filename>)`.
Default captures first two digits.}
}
\value{
Invisible list of new file paths.
}
\description{
Moves mzML files into plex-specific subdirectories based on filename regex.
}
\examples{
\dontrun{
organize_mzml_by_plex("E:/Data_E/Tyler/5306_global/")
}
}
