% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pecora_analysis.R
\name{pecora_analysis}
\alias{pecora_analysis}
\title{Performs the PeCorA analysis described in https://www.ncbi.nlm.nih.gov/pmc/articles/PMC8592057/.
Take a look at the linked paper to get an idea of the procedure.}
\usage{
pecora_analysis(
  m,
  treatment_string,
  control_group = NULL,
  proteins,
  sample_standardize = TRUE,
  peptide_standardize = TRUE,
  folder = "",
  save_plots = TRUE,
  median_mod = FALSE
)
}
\arguments{
\item{m}{msnset object containing a pData AND fData table. fData must have a Protein and Peptide column.}

\item{treatment_string}{the name of the column in pData containing treatment information. The data can be numerical of categorical.}

\item{proteins}{character vector indicating which proteins to analyze.}

\item{sample_standardize}{logical for whether exprs data should be standardized so that sample wise we have mean zero and variance one.}

\item{peptide_standardize}{logical for whether exprs data should be normalized so that logratio data is relative to the control group.}

\item{folder}{folder in which to save the plots. For example folder = "directory/".}

\item{save_plots}{logical for whether to save boxplots/scatterplots of the significant peptides.}
}
\value{
PeCorA results table for the supplied proteins. Saves plots of the significant peptides to the given folder.
}
\description{
Takes an msnset, a list of proteins, and a treatment variable. The data is standardized depending on the
chosen parameters of 'sample_standardize' and 'peptide_standardize'. Then the PeCorA function from the
PeCorA package is used to find discordant peptides. Plots of the significant peptides are made and
saved to the given folder. Finally, the output contains a table with the phosphosites mapping to the given
proteins, along with their PeCorA analysis p-values.
}
