% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sel_feat_with_p_vals.R
\name{plot_sel_feat_with_p_vals}
\alias{plot_sel_feat_with_p_vals}
\title{MSnSet Feature Bar Chart with (Adjusted) P-Value Coloring}
\usage{
plot_sel_feat_with_p_vals(
  top_selected,
  msnset,
  response_colname,
  feat_name_colname = NULL,
  highlight_feats = NULL,
  highlight_reason = NULL,
  alpha = 0.05,
  title = "Frequency of\\nBoruta-Selected Features"
)
}
\arguments{
\item{top_selected}{A named vector with names as top Boruta-selected features and \
values as the frequency at which said features were selected}

\item{msnset}{A \code{MSnSet} object that represented the input to an `rf_modeling` \
call (with Boruta turned on). This is required so P-values can be computed}

\item{response_colname}{The column name (in \code{pData(msnset)}) of the machine \
learning response variable}

\item{feat_name_colname}{If provided, a column in \code{fData(msnset)} by which \
to name the features by. If \code{NULL}, then \
\code{featureNames(msnset) will be used.}}

\item{highlight_feats}{If provided, put an asterisk ❉ next to the label of these \
features. Note: this uses \code{feat_name_colnames} if passed, and not \
\code{msnset}'s native features.}

\item{highlight_reason}{Short string describing why certain features would be \
highlighted. Only applicable if \code{highlight_feats} is not \code{NULL}.}

\item{alpha}{The alpha-level that defines statistical significance (for purposes \
of the color scale)}
}
\description{
MSnSet Feature Bar Chart with (Adjusted) P-Value Coloring
}
