% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_shap.R
\name{plot_treeshap}
\alias{plot_treeshap}
\title{Plot SHAP values for Random Forest model}
\usage{
plot_treeshap(
  msnset,
  features_or_modeling_out,
  response,
  pred.cls,
  interpretation_and_direction = 1
)
}
\arguments{
\item{msnset}{MSnSet object containing the data for analysis.}

\item{features_or_modeling_out}{Either the output of a call to functions like
\code{rf_modeling} (which contains an element \code{top}) OR
a character vector of feature names to be used for SHAP value calculation.
The former will be assumed if \code{features_or_modeling_out} has names
 "prob", "features", "top", "auc", "pred".}

\item{response}{Character. The name of the response variable in \code{pData(msnset)}.}

\item{pred.cls}{Character. The level of the response variable that corresponds to the positive class (the one being predicted).}

\item{interpretation_and_direction}{Integer. Either 1 or -1. Direction of interpretation.
Default is 1. If 1, the interpretation is that feature abundances are indeed abundances.
If -1, the interpretation is that feature abundances represent changes in abundance
(e.g., log2 fold changes between conditions). This is for historical purposes. If this is not desired,
you may change axis labels by modifying the returned ggplot object.}
}
\value{
A ggplot object showing SHAP values and protein abundances.
}
\description{
This function calculates and visualizes SHAP (SHapley Additive exPlanations) values
for a *Random Forest* model trained on an MSnSet object. It shows the impact of
each feature on the model's prediction of a specific class, along with the
underlying protein abundance.
}
\references{
SHAP paper:
  doi = 10.5555/3295222.3295230,
  author = Lundberg, Scott M. and Lee, Su-In,
  title = A unified approach to interpreting model predictions,
  year = 2017,
  isbn = 9781510860964,
  publisher = Curran Associates Inc.,
  address = Red Hook, NY, USA,
  abstract = Understanding why a model makes a certain prediction can be as crucial as the prediction's accuracy in many applications. However, the highest accuracy for large modern datasets is often achieved by complex models that even experts struggle to interpret, such as ensemble or deep learning models, creating a tension between accuracy and interpretability. In response, various methods have recently been proposed to help users interpret the predictions of complex models, but it is often unclear how these methods are related and when one method is preferable over another. To address this problem, we present a unified framework for interpreting predictions, SHAP (SHapley Additive exPlanations). SHAP assigns each feature an importance value for a particular prediction. Its novel components include: (1) the identification of a new class of additive feature importance measures, and (2) theoretical results showing there is a unique solution in this class with a set of desirable properties. The new class unifies six existing methods, notable because several recent methods in the class lack the proposed desirable properties. Based on insights from this unification, we present new methods that show improved computational performance and/or better consistency with human intuition than previous approaches.,
  booktitle = Proceedings of the 31st International Conference on Neural Information Processing Systems,
  pages = 4768–4777,
  numpages = 10,
  location = Long Beach, California, USA,
  series = NIPS'17
}
