% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAUC.R
\name{plotAUC}
\alias{plotAUC}
\title{Plot AUC}
\usage{
plotAUC(modelingResult, CI = FALSE, ...)
}
\arguments{
\item{modelingResult}{output of \code{lr_modeling} or \code{rf_modeling}.}

\item{CI}{(logical) whether to plot confidence intervals. Default is
\code{FALSE}.}

\item{...}{further arguments passed to the \code{plot} method of the
\code{\link[pROC]{ci.se}} object of the \code{pROC} package.}
}
\description{
Plot AUC after LOOCV model evaluation.
}
\seealso{
\code{\link{plotAUC_gg}}
}
