% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC_report.R
\name{qc_report}
\alias{qc_report}
\title{Generate QC Report}
\usage{
qc_report(
  msnset,
  msnset_orig,
  p_data_contrasts,
  the_sch_samples,
  report_output_path,
  plot_cache_dir = NULL,
  report_title = "QC Report",
  self_contained = TRUE,
  feature_kind = "Feature",
  img_dims = list(boxplot = c(8, 6), feature_count = c(8, 6), missingness = c(8, 6),
    multi_pca = c(8, 12))
)
}
\arguments{
\item{msnset}{MSnSet. The processed version (e.g., after normalization if desired).}

\item{msnset_orig}{MSnSet. The original object (for missingness plots).}

\item{p_data_contrasts}{Character. Vector of pData columns to use as
primary and secondary grouping variables}

\item{the_sch_samples}{Character. Vector of sample names for the detailed
correlation heatmap. If NULL and that plot is requested, an error will be raised.}

\item{report_output_path}{Character. File path where the report (\code{.qmd})
will be saved.}

\item{plot_cache_dir}{Character. Optional. Directory to cache plot objects
as \code{.rds} files.}

\item{report_title}{Character. Title of the report.}

\item{self_contained}{Logical. Whether the HTML report should be self-contained.}

\item{feature_kind}{Character. Label for features (e.g., "Protein"). Used for legend labeling.}

\item{img_dims}{List. Named list of numeric vectors defining \code{c(width, height)}
for each plot type. This is the primary interface for controlling which plots are generated.
Supported names are: \code{boxplot}, \code{feature_count}, \code{missingness}, \code{multi_pca},
\code{overall_correlation_heatmap}, and \code{samples_correlation_heatmap}.}
}
\value{
None. Saves a file to \code{report_output_path} and renders it.
}
\description{
Streamlines the creation of multiple QC visualizations and compiles them
into a Quarto/Markdown report.
}
\details{
ADDINS.PNNL is required.
}
