% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_FragPipe_LFQ.R
\name{read_FragPipe_LFQ}
\alias{read_FragPipe_LFQ}
\title{Reading MSFragger-generated LFQ-based MSstats from a file path as MSnSet object}
\usage{
read_FragPipe_LFQ(path = NULL)
}
\arguments{
\item{path}{character; File path to the FragPipe-generated MSstats.csv file}
}
\value{
(MSnSet) MSnSet object of MSFragger LFQ results
}
\description{
Function has only been tested with label-free intensity-based
  quantification data. MSstats.csv is
  an optional output file which needs to be specified in FP settings.
}
\examples{
file_path <- "C:/Users/fakeusr222/Desktop/MSF_LFQ_job/MSstats.csv"
  msnset <- read_FragPipe_LFQ(file_path)
  show(msnset)

}
