% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_FragPipe_TMT.R
\name{read_FragPipe_TMT}
\alias{read_FragPipe_TMT}
\title{Reading MSFragger-generated tmt-report files from a file path as MSnSet object}
\usage{
read_FragPipe_TMT(
  path = NULL,
  org_to_retain = NULL,
  use_gene_as_prot_id = TRUE
)
}
\arguments{
\item{path}{character; File path to the desired FragPipe-generated tmt-report file.
Any tmt-report file may be used.}

\item{use_gene_as_prot_id}{logical; Used only in case of `single-site` files. Switches
notation from UniProt_Site to a more human-readable and conventional Gene-Site.
Default it `TRUE`. In case there are duplicates in the new site IDs, returns
the error message with the prompt to switch to `FALSE`.}
}
\value{
(MSnSet) MSnSet object of MSFragger TMT results
}
\description{
Function has only been tested with TMT intensity-based
  quantification data. Desired tmt-report output files (e.g., "ratio_multi-site_MD.tsv")
  must be properly selected in the FP settings.
}
\examples{
file_path <- "C:/Users/fakeusr222/Desktop/MSF_TMT_job/ratio_multi-site_MD.tsv"
  msnset <- read_FragPipe_TMT(file_path)
  show(msnset)

}
