% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_by_size_enrichments.R
\name{subset_by_size}
\alias{subset_by_size}
\title{Filter of ontology enrichment results by size}
\usage{
subset_by_size(
  x,
  minObsSize = 0,
  maxObsSize = Inf,
  minSetSize = 0,
  maxSetSize = Inf
)
}
\arguments{
\item{x}{is an \code{enrichResults} instance}

\item{minObsSize}{minimum number of genes must be observed per set}

\item{maxObsSize}{maximum number of genes must be observed per set}

\item{minSetSize}{minimum number of genes must be in the set}

\item{maxSetSize}{maximum number of genes must be in the set}
}
\value{
filtered \code{enrichResults} instance
}
\description{
This utility works to filter results of enrich* test
from \code{\link[clusterProfiler]{clusterProfiler-package}},
\code{\link[enrichit]{enrichit-package}} or
\code{\link[ReactomePA]{ReactomePA-package}}
}
\examples{
library(clusterProfiler)
library(enrichplot)
library(org.Hs.eg.db)
data(gcSample)
x <- enrichGO(gcSample[[6]],
              universe = unique(unlist((gcSample))),
              OrgDb = "org.Hs.eg.db",
              ont = "CC",
              pvalueCutoff = 1,
              qvalueCutoff = 1,
              minGSSize = 0)
res <- x@result
res$geneID <- NULL
rownames(res) <- NULL
head(res)
cnetplot(x)

# E.g. the "cell", "cell part" & "cellular_component" seem too generic
x <- subset_by_size(x, maxObsSize=50)
res <- x@result
res$geneID <- NULL
rownames(res) <- NULL
# Now they are gone
head(res)
cnetplot(x)
}
