% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptm_adjustment_by_global.R
\name{adjust_for_parent_protein_profiles}
\alias{adjust_for_parent_protein_profiles}
\title{Adjusting PTM profiles by global}
\usage{
adjust_for_parent_protein_profiles(m_ptm, m_global)
}
\arguments{
\item{m_ptm}{MSnSet with PTM data}

\item{m_global}{MSnSet with global data}
}
\value{
MSnSet object
}
\description{
Adjusting PTM relative abundance profiles by the global profiles of the
parent protein. The motivation behind this adjustment is that there are two
contributions to the observed PTM abundance changes - change in the
PTM stoichiometry and the change in the parent protein abundance.
}
