% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_stats.R
\name{compute_enrichment_stats}
\alias{compute_enrichment_stats}
\title{Calculating basic enrichment statistics.}
\usage{
compute_enrichment_stats(enriched, target, background)
}
\arguments{
\item{enriched}{list of enriched proteins}

\item{target}{list of proteins targeted for enrichment}

\item{background}{list of background proteins}
}
\value{
numeric
}
\description{
Type of identification does not matter as long as it is the same -- be
that gene symbols or UniProt accessions. Computes fold enrichment and
Matthews Correlation Coefficient (MCC), which unlike fold enrichment
also accounts for true negatives and is more robust to class imbalance.
Absolute values depend on the choice of `background`; comparisons
across enrichment lists should use the same `background`. Ideally, the
`background` should be the detectable proteome. However, if the detectable
proteome (e.g. from global samples) is unavailable, then it is OK to use
the entire genome. Use of entire genome will make the numbers biased, since
unexpressed proteins will inflate true negatives, but it is OK to use for
relative comparison of the enrichment methods.
}
\details{
Calculates fold enrichment and Matthews Correlation Coefficient
}
