% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_processing.R
\name{create_msnset}
\alias{create_msnset}
\title{Create MSnSet Object from Expression Data and FASTA}
\usage{
create_msnset(
  path_to_exprs_or_exprs_df,
  path_to_fasta_for_f_data,
  p_data_df,
  feature_name_col = c("ensembl_protein", "ensembl_gene"),
  semicolon_warning = TRUE,
  subset_samples_by_p_data = TRUE,
  save_as_rds_path = NULL
)
}
\arguments{
\item{path_to_exprs_or_exprs_df}{Character or data.frame. Path to a file containing expression levels or a data.frame of expression levels.}

\item{path_to_fasta_for_f_data}{Character. Path to the FASTA file used for feature data.}

\item{p_data_df}{data.frame. Metadata with row names matching sample names in the expression data.}

\item{feature_name_col}{Character. Column name in the FASTA-derived feature data to use for matching with the \code{Index} column.
Should be one of \code{"ensembl_protein"} or \code{"ensembl_gene"}.}

\item{semicolon_warning}{Logical. Whether to warn if IDs contain semicolons. Default is TRUE.}

\item{subset_samples_by_p_data}{Logical. Whether to subset samples in the expression levels by those present in \code{p_data_df}. Default is \code{TRUE}.}

\item{save_as_rds_path}{Character. Optional path to save the final \code{MSnSet} object as an \code{.rds} file. Default is NULL.}
}
\value{
An MSnSet object containing the processed data.
}
\description{
Generates an MSnSet object using expression levels and a FASTA file for feature data.
}
\details{
\strong{This function has not been extensively tested with DIA data.}
\strong{This function is currently somewhat brittle, as it requires a specific TMT format
from FragPipe and a specific FASTA header format (which is from Ensembl).}

\strong{Sample Alignment between expression data and pData:}
The function aligns the expression levels with sample metadata
(\code{p_data_df}). It identifies the column range for samples by looking for landmark columns
like \code{ReferenceIntensity} (typical for TMT from FragPipe) or \code{Transcript.Id} in the case of DIA data.
This function subsets both the data and the metadata to include only the samples
found in both. In this case, a warning is issued.

\strong{Feature Alignment between expression data and fData:}
This function pulls in feature-level metadata from a
FASTA file. It parses standard sequence headers to extract Ensembl identifiers (Protein,
Transcript, and Gene IDs) and descriptions. This function subsets the expression levels to
include only features that can be mapped to the FASTA entries. In this case, a warning is issued.

\itemize{
  \item \strong{global proteomics:} For standard global studies, it uses the protein or
  gene index as a simple unique identifier.
  \item \strong{PTM proteomics:} For modification-site-level data,
  it detects the presence of a \code{Peptide} column and automatically constructs unique
  identities for each site in the format \code{Index-site@Peptide}
}

This function resolves potential duplicates (e.g., from multiple peptides
mapping to one protein) by prioritizing the identification with the highest confidence
(\code{MaxPepProb}).
}
