% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_shap.R
\name{plot_treeshap}
\alias{plot_treeshap}
\title{Plot SHAP values for Random Forest model}
\usage{
plot_treeshap(
  msnset,
  features_or_modeling_out,
  response,
  pred.cls,
  interpretation_and_direction = 1
)
}
\arguments{
\item{msnset}{MSnSet object containing the data for analysis.}

\item{features_or_modeling_out}{Either the output of a call to functions like
\code{rf_modeling} (which contains an element \code{top}) OR
a character vector of feature names to be used for SHAP value calculation.
The former will be assumed if \code{features_or_modeling_out} has names
 "prob", "features", "top", "auc", "pred".}

\item{response}{Character. The name of the response variable in \code{pData(msnset)}.}

\item{pred.cls}{Character. The level of the response variable that corresponds to the positive class (the one being predicted).}

\item{interpretation_and_direction}{Integer. Either 1 or -1. Direction of interpretation.
Default is 1. If 1, the interpretation is that feature abundances are indeed abundances.
If -1, the interpretation is that feature abundances represent changes in abundance
(e.g., log2 fold changes between conditions). This is for historical purposes. If this is not desired,
you may change axis labels by modifying the returned ggplot object.}
}
\value{
A ggplot object showing SHAP values and protein abundances.
}
\description{
This function calculates and visualizes SHAP (SHapley Additive exPlanations) values
for a *Random Forest* model trained on an MSnSet object. It shows the impact of
each feature on the model's prediction of a specific class, along with the
underlying protein abundance.
}
\references{
Lundberg, S. M., & Lee, S. I. (2017). A unified approach to interpreting model predictions.
Advances in neural information processing systems, 30.
\url{https://proceedings.neurips.cc/paper/2017/hash/8a20a8621978632d76c43dfd28b67767-Abstract.html}
}
