% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_processing.R
\name{run_msConvert_cli}
\alias{run_msConvert_cli}
\title{Run msConvert CLI to Convert Raw Files to mzML}
\usage{
run_msConvert_cli(
  raw_files_dir,
  output_dir,
  msconvert_path = NULL,
  addl_cli_args = c("--zlib", "--filter", "\\"peakPicking vendor msLevel=1-\\"",
    "--filter", "\\"titleMaker", "<RunId>.<ScanNumber>.<ScanNumber>.<ChargeState>",
    "File:\\"\\"\\"^<SourcePath^>\\"\\"\\", NativeID:\\"\\"\\"^<Id^>\\"\\"\\"\\""),
  overwrite = FALSE
)
}
\arguments{
\item{raw_files_dir}{Character. Directory containing .raw files to convert.}

\item{output_dir}{Character. Directory to save converted .mzML files.}

\item{msconvert_path}{Character. Path to msconvert executable. If NULL, will use msconvert from system PATH.}

\item{addl_cli_args}{Character vector. Additional command-line arguments for msconvert. Default includes zlib compression and peak picking.}

\item{overwrite}{Logical. Whether to overwrite existing files. Default is FALSE.}
}
\value{
Invisible list of msconvert command outputs for each file.
}
\description{
Converts Thermo .raw files to .mzML format using the msConvert command-line tool.
}
\examples{
\dontrun{
run_msConvert_cli(
   raw_files_dir = "E:/Data_E/Tyler/5306_global/raw",
   output_dir = "E:/Data_E/Tyler/5306_global/mzML"
)
}
}
